// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ROLE = {
  customer: 'customer',
  provider: 'provider',
  admin: 'admin'
} as const

export type ROLE = (typeof ROLE)[keyof typeof ROLE]


export const USERSTATUS = {
  suspend: 'suspend',
  activate: 'activate'
} as const

export type USERSTATUS = (typeof USERSTATUS)[keyof typeof USERSTATUS]


export const ORDERSTATUS = {
  PLACED: 'PLACED',
  PREPARING: 'PREPARING',
  READY: 'READY',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED'
} as const

export type ORDERSTATUS = (typeof ORDERSTATUS)[keyof typeof ORDERSTATUS]